[BE-Feat] 논의 확정 이후 비즈니스 로직 구현 (#119)
* feat: deleteSharedEvent() 에러처리 추가

* refactor: SharedEventRequest의 파라미터를 Api 명세에 맞추어 수정

* chore: getShareEvent에 readOnly 설정

* test: SharedEventServiceTest에 테스트 코드 추가

* feat: discussionId pathVariable에 대해 Validation 추가

* feat: ApiException 에서 super()로 메세지 반영

* feat: UserProfile Dto 추가

* feat: DiscussionParticipantRepository에 메서드 추가

1. findUserProfilesByDiscussionId()추가

2. DiscussionService.confirmSchedule()에서 해당 메서드 사용하도록 변경

* refactor: SharedEventResponse를 SharedEventDto로 이름 변경

* refactor: UserProfile 삭제, 대신 User 엔티티를 사용하도록 변경

* feat: 비동기 사용을 위해 EnableAsyncConfig 추가

* feat: 논의 확정시 참여자들의 개인 캘린더에 공유 일정 추가하는 로직 구현

* feat: 데이터 삭제 메서드를 비동기로 변경

* feat: 논의 확정 시 Redis 데이터 삭제 로직 구현

* feat: DiscussionParticipantService 생성 후 관련 로직 위임

* refactor: createPersonalEventForParticipants()를 stream 사용하도록 변경

* feat: readOnly 추가

* feat: findUsersByDiscussionId를 fetch join으로 변경

* refactor: createPersonalEventsForParticipants를 동기로 변경

* refactor: redis 데이터 삭제를 최대 3번까지 재시도 하도록 수정

+++ backend/src/main/java/endolphin/backend/domain/shared_event/SharedEventService.java
@@ -2,7 +2,7 @@ package endolphin.backend.domain.shared_event;
 
 import endolphin.backend.domain.discussion.entity.Discussion;
 import endolphin.backend.domain.shared_event.dto.SharedEventRequest;
-import endolphin.backend.domain.shared_event.dto.SharedEventResponse;
+import endolphin.backend.domain.shared_event.dto.SharedEventDto;
 import endolphin.backend.domain.shared_event.entity.SharedEvent;
 import endolphin.backend.global.error.exception.ApiException;
 import endolphin.backend.global.error.exception.ErrorCode;
@@ -18,27 +18,31 @@ public class SharedEventService {
 
     private final SharedEventRepository sharedEventRepository;
 
-    public SharedEventResponse createSharedEvent(Discussion discussion,
+    public SharedEventDto createSharedEvent(Discussion discussion,
         SharedEventRequest request) {
-        Validator.validateDateTimeRange(request.startTime(), request.endTime());
+        Validator.validateDateTimeRange(request.startDateTime(), request.endDateTime());
 
         SharedEvent sharedEvent = SharedEvent.builder()
             .discussion(discussion)
-            .start(request.startTime())
-            .end(request.endTime())
+            .start(request.startDateTime())
+            .end(request.endDateTime())
             .build();
 
-        return SharedEventResponse.of(sharedEventRepository.save(sharedEvent));
+        return SharedEventDto.of(sharedEventRepository.save(sharedEvent));
     }
 
-    public SharedEventResponse getSharedEvent(Long sharedEventId) {
+    @Transactional(readOnly = true)
+    public SharedEventDto getSharedEvent(Long sharedEventId) {
         SharedEvent sharedEvent = sharedEventRepository.findById(sharedEventId)
             .orElseThrow(() -> new ApiException(ErrorCode.SHARED_EVENT_NOT_FOUND));
 
-        return SharedEventResponse.of(sharedEvent);
+        return SharedEventDto.of(sharedEvent);
     }
 
     public void deleteSharedEvent(Long sharedEventId) {
+        if (!sharedEventRepository.existsById(sharedEventId)) {
+            throw new ApiException(ErrorCode.SHARED_EVENT_NOT_FOUND);
+        }
         sharedEventRepository.deleteById(sharedEventId);
     }
